+++ src/main.ts
@@ -1,5 +1,5 @@
-import { bgRed } from "colorette";
-import type { CommandDef } from "./types";
+import consola from "consola";
+import type { ArgsDef, CommandDef } from "./types";
 import { resolveSubCommand, runCommand } from "./command";
 import { CLIError } from "./_utils";
 import { showUsage } from "./usage";
@@ -8,26 +8,40 @@ export interface RunMainOptions {
   rawArgs?: string[];
 }
 
-export async function runMain(cmd: CommandDef, opts: RunMainOptions = {}) {
+export async function runMain<T extends ArgsDef = ArgsDef>(
+  cmd: CommandDef<T>,
+  opts: RunMainOptions = {},
+) {
   const rawArgs = opts.rawArgs || process.argv.slice(2);
   try {
     if (rawArgs.includes("--help") || rawArgs.includes("-h")) {
       await showUsage(...(await resolveSubCommand(cmd, rawArgs)));
       process.exit(0);
+    } else if (rawArgs.length === 1 && rawArgs[0] === "--version") {
+      const meta =
+        typeof cmd.meta === "function" ? await cmd.meta() : await cmd.meta;
+      if (!meta?.version) {
+        throw new CLIError("No version specified", "E_NO_VERSION");
+      }
+      consola.log(meta.version);
     } else {
       await runCommand(cmd, { rawArgs });
     }
   } catch (error: any) {
     const isCLIError = error instanceof CLIError;
     if (!isCLIError) {
-      console.error(error, "\n");
+      consola.error(error, "\n");
     }
-    console.error(
-      `\n${bgRed(` ${error.code || error.name} `)} ${error.message}\n`
-    );
     if (isCLIError) {
       await showUsage(...(await resolveSubCommand(cmd, rawArgs)));
     }
+    consola.error(error.message);
     process.exit(1);
   }
+}
+
+export function createMain<T extends ArgsDef = ArgsDef>(
+  cmd: CommandDef<T>,
+): (opts?: RunMainOptions) => Promise<void> {
+  return (opts: RunMainOptions = {}) => runMain(cmd, opts);
 }