+++ src/usage.ts
@@ -1,16 +1,24 @@
+import consola from "consola";
+import { colors } from "consola/utils";
 import { formatLineColumns, resolveValue } from "./_utils";
-import type { CommandDef } from "./types";
+import type { ArgsDef, CommandDef } from "./types";
 import { resolveArgs } from "./args";
 
-export async function showUsage(cmd: CommandDef, parent?: CommandDef) {
+export async function showUsage<T extends ArgsDef = ArgsDef>(
+  cmd: CommandDef<T>,
+  parent?: CommandDef<T>,
+) {
   try {
-    console.log((await renderUsage(cmd, parent)) + "\n");
+    consola.log((await renderUsage(cmd, parent)) + "\n");
   } catch (error) {
-    console.error(error);
+    consola.error(error);
   }
 }
 
-export async function renderUsage(cmd: CommandDef, parent?: CommandDef) {
+export async function renderUsage<T extends ArgsDef = ArgsDef>(
+  cmd: CommandDef<T>,
+  parent?: CommandDef<T>,
+) {
   const cmdMeta = await resolveValue(cmd.meta || {});
   const cmdArgs = resolveArgs(await resolveValue(cmd.args || {}));
   const parentMeta = await resolveValue(parent?.meta || {});
@@ -30,7 +38,7 @@ export async function renderUsage(cmd: CommandDef, parent?: CommandDef) {
       const isRequired = arg.required !== false && arg.default === undefined;
       // (isRequired ? " (required)" : " (optional)"
       const usageHint = arg.default ? `="${arg.default}"` : "";
-      posLines.push([name + usageHint, arg.description || ""]);
+      posLines.push(["`" + name + usageHint + "`", arg.description || ""]);
       usageLine.push(isRequired ? `<${name}>` : `[${name}]`);
     } else {
       const isRequired = arg.required === true && arg.default === undefined;
@@ -41,7 +49,7 @@ export async function renderUsage(cmd: CommandDef, parent?: CommandDef) {
               `--no-${arg.name}`,
             ].join(", ")
           : [...(arg.alias || []).map((a) => `-${a}`), `--${arg.name}`].join(
-              ", "
+              ", ",
             )) +
         (arg.type === "string" && (arg.valueHint || arg.default)
           ? `=${
@@ -53,7 +61,7 @@ export async function renderUsage(cmd: CommandDef, parent?: CommandDef) {
         description = arg.negativeDescription
       }
       argLines.push([
-        argStr + (isRequired ? " (required)" : ""),
+        "`" + argStr + (isRequired ? " (required)" : "") + "`",
         description || "",
       ]);
       if (isRequired) {
@@ -64,8 +72,11 @@ export async function renderUsage(cmd: CommandDef, parent?: CommandDef) {
 
   if (cmd.subCommands) {
     const commandNames: string[] = [];
-    for (const [name, sub] of Object.entries(cmd.subCommands)) {
-      commandsLines.push([name, sub.meta?.description || ""]);
+    const subCommands = await resolveValue(cmd.subCommands);
+    for (const [name, sub] of Object.entries(subCommands)) {
+      const subCmd = await resolveValue(sub);
+      const meta = await resolveValue(subCmd?.meta);
+      commandsLines.push([`\`${name}\``, meta?.description || ""]);
       commandNames.push(name);
     }
     usageLine.push(commandNames.join("|"));
@@ -74,38 +85,42 @@ export async function renderUsage(cmd: CommandDef, parent?: CommandDef) {
   const usageLines: (string | undefined)[] = [];
 
   const version = cmdMeta.version || parentMeta.version;
+
   usageLines.push(
-    commandName + (version ? ` v${version}` : ""),
-    cmdMeta.description,
-    ""
+    colors.gray(
+      `${cmdMeta.description} (${
+        commandName + (version ? ` v${version}` : "")
+      })`,
+    ),
+    "",
   );
 
   const hasOptions = argLines.length > 0 || posLines.length > 0;
   usageLines.push(
-    `USAGE: ${commandName}${hasOptions ? " [OPTIONS]" : ""} ${usageLine.join(
-      " "
-    )}`,
-    ""
+    `${colors.underline(colors.bold("USAGE"))} \`${commandName}${
+      hasOptions ? " [OPTIONS]" : ""
+    } ${usageLine.join(" ")}\``,
+    "",
   );
 
   if (posLines.length > 0) {
-    usageLines.push("ARGUMENTS:", "");
+    usageLines.push(colors.underline(colors.bold("ARGUMENTS")), "");
     usageLines.push(formatLineColumns(posLines, "  "));
     usageLines.push("");
   }
 
   if (argLines.length > 0) {
-    usageLines.push("OPTIONS:", "");
+    usageLines.push(colors.underline(colors.bold("OPTIONS")), "");
     usageLines.push(formatLineColumns(argLines, "  "));
     usageLines.push("");
   }
 
   if (commandsLines.length > 0) {
-    usageLines.push("COMMANDS:", "");
+    usageLines.push(colors.underline(colors.bold("COMMANDS")), "");
     usageLines.push(formatLineColumns(commandsLines, "  "));
     usageLines.push(
       "",
-      `Use \`${commandName} <command> --help\` for more information about a command.`
+      `Use \`${commandName} <command> --help\` for more information about a command.`,
     );
   }